// src/interfaces/ResponseUser.ts
interface UserResponse {
  idUser: number;
  rutUser: string;
  phoneUser: string;
  emailUser: string;
  birthdayUser: string;
  nameUser: string;
  firstLastNameUser: string;
  secondLastNameUser: string;
  passwordUser: string;
  statusUser: string;
  typeUser: string;
  idAddress: number;
}

export default UserResponse;